web-of-soil: added sections to web-of-soil. interlocked slides from different sections. moved media hub into a component and created useMediaHub hook
